Fixed some broken things

+++ src/main/java/com/example/Appointment/repository/TeacherDataRepository.java
@@ -3,5 +3,5 @@ package com.example.Appointment.repository;
 import com.example.Appointment.entity.TeacherData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TeacherDataRepository extends JpaRepository<TeacherData, Long> {
+public interface TeacherDataRepository extends JpaRepository<TeacherData, Integer> {
 }
